+++ src/form/RruMultiSelectInput/types/RruMultiSelectInputProps.ts
@@ -40,4 +40,6 @@ export default interface RruMultiSelectInputProps {
    * Controls whether the input field is disabled or not
    */
   disabled?: boolean;
+
+  placeholder?: string;
 }